Start CDO FOAF shapes from base branch
Note: The retrieval of foaf.rdf from archive.org is in response to this
alert on the FOAF home page:

> ALERT! We are currently serving an older version of the FOAF spec,
> from 2004, while restoring from archives. Although it did not change
> radically between 2004 and 2014 (last release), please consult the
> Internet archive ( html, rdf) for that and earlier releases, which
> extend back to our earliest version in 2000.

This retrieval is as was done in UCO-Profile-FOAF commit:
8d516e41ae73c5a8e6842450154f4a4056691b7e

Other files in this commit were also ported from that same UCO FOAF
Profile commit, as part of transitioning FOAF-generic shapes from the
UCO Profile.

A follow-on patch will generate Make-managed files.

References:
* http://xmlns.com/foaf/0.1/
* https://github.com/ucoProject/UCO-Profile-FOAF

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>